+++ src/api/app.js
@@ -2,7 +2,12 @@ const express = require('express');
 
 const routes = require('./routes');
 
+require('../database');
+
 const app = express();
+
+app.use(express.json());
+app.use(routes);
 
 // Não remover esse end-point, ele é necessário para o avaliador
 app.get('/', (request, response) => {
@@ -10,8 +15,4 @@ app.get('/', (request, response) => {
 });
 // Não remover esse end-point, ele é necessário para o avaliador
 
-app.use(express.json());
-app.use(routes);
-
 module.exports = app;
-// export default new App().server;
